Split unified agent grouping helpers
